% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting_tmb.R
\name{fit_zoidTMB}
\alias{fit_zoidTMB}
\title{Fit a trinomial mixture model with TMB}
\usage{
fit_zoidTMB(
  formula = NULL,
  design_matrix,
  data_matrix,
  overdispersion = FALSE,
  overdispersion_sd = 5,
  prior_sd = NA
)
}
\arguments{
\item{formula}{The model formula for the design matrix. Does not need to have a response specified. If =NULL, then
the design matrix is ignored and all rows are treated as replicates}

\item{design_matrix}{A data frame, dimensioned as number of observations, and covariates in columns}

\item{data_matrix}{A matrix, with observations on rows and number of groups across columns}

\item{overdispersion}{Whether or not to include overdispersion parameter, defaults to FALSE}

\item{overdispersion_sd}{Prior standard deviation on 1/overdispersion parameter, Defaults to inv-Cauchy(0,5)}

\item{prior_sd}{Optional prior sd / penalty for fixed effects}
}
\description{
Fit a trinomial mixture model that optionally includes covariates to estimate
effects of factor or continuous variables on proportions.
}
\examples{
\donttest{
y <- matrix(c(3.77, 6.63, 2.60, 0.9, 1.44, 0.66, 2.10, 3.57, 1.33),
  nrow = 3, byrow = TRUE
)
# fit a model with no covariates
fit <- fit_zoidTMB(data_matrix = y)

# fit a model with 1 factor
design <- data.frame("fac" = c("spring", "spring", "fall"))
fit <- fit_zoidTMB(formula = ~fac, design_matrix = design, data_matrix = y)

# try a model with random effects
set.seed(123)
y <- matrix(runif(99,1,4), ncol=3)
design <- data.frame("fac" = sample(letters[1:5], size=nrow(y), replace=TRUE))
design$fac <- as.factor(design$fac)
fit <- fit_zoid(formula = ~(1|fac), design_matrix = design, data_matrix = y)
}

}
